requirements.txt
PyMuPDF>=1.23.0
google-generativeai>=0.3.0
flask>=3.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
Jinja2>=3.1.0
